feat: se añadio el componente LoginButton en el navbar
