+++ frontend/src/components/sections/Welcome/MainBox/MainBox.tsx
@@ -1,17 +1,8 @@
 import { Box, Typography } from "@mui/material";
-import { CustomButton } from "@ui/CustomButton/CustomButton";
-import { useModalStore } from "@store/modal.store";
+import { LoginButton } from "./LoginButton";
 import { MainBoxStyles } from "./MainBoxStyles";
 
 export const MainBox = () => {
-    const setModalData = useModalStore((state) => state.setModalData);
-    const handleOpenModal = () => {
-        setModalData({
-            showModal: true,
-            title: "¿Querés agendar una cita?",
-            operation: "menu"
-        });
-    }
     return (
         <Box sx={MainBoxStyles.mainBox}>
             <Box sx={MainBoxStyles.welcomeBox}>
@@ -25,9 +16,7 @@ export const MainBox = () => {
                         Tu lugar para pedir un turno
                     </Typography>
                 </Box>
-                <CustomButton onClick={handleOpenModal}>
-                    <Typography fontFamily="Inria Sans Bold" textTransform="none">Agenda tu cita</Typography>
-                </CustomButton>
+                <LoginButton />
             </Box>
         </Box>
     );
